Refactor query test cases to share code
