+++ tests/test_postgres.py
@@ -10,6 +10,7 @@ import pytest
 from pytest_cases import parametrize_with_cases
 
 from .test_cases import drop_tables_cases as dtc
+from .test_cases import query_cases as qc
 
 if TYPE_CHECKING:
     from unittest.mock import MagicMock
@@ -62,4 +63,59 @@ def test_drop_tables(
                 WHERE table_schema='public'
                 """,
         )
-        assert sorted([r[0] for r in res.fetchall()]) == sorted(tc.expected_values)
+        assert sorted([r[0] for r in res.fetchall()]) == sorted(tc.expected_tables)
+
+
+@mock.patch("ldlite.request_get")
+@parametrize_with_cases("tc", cases=qc.QueryTestCases)
+def test_query(
+    request_get_mock: MagicMock,
+    pg_dsn: None | Callable[[str], str],
+    tc: qc.QueryCase,
+) -> None:
+    if pg_dsn is None:
+        pytest.skip("Specify the pg host using --pg-host to run")
+
+    from ldlite import LDLite as uut
+
+    ld = uut()
+    tc.patch_request_get(ld, request_get_mock)
+    dsn = pg_dsn(tc.db)
+    ld.connect_db_postgresql(dsn)
+
+    for prefix in tc.values:
+        ld.query(table=prefix, path="/patched", json_depth=tc.json_depth)
+
+    catalog = f"{prefix}__tcatalog"
+    fixed_cols = [prefix]
+    # TODO: Clean this edge case up so tcatalog is always created
+    if tc.json_depth > 0:
+        fixed_cols.append(catalog)
+
+    with psycopg2.connect(dsn) as conn:
+        with conn.cursor() as res:
+            res.execute(
+                """
+                SELECT table_name
+                FROM information_schema.tables
+                WHERE table_schema='public'
+                """,
+            )
+            assert sorted([r[0] for r in res.fetchall()]) == sorted(
+                [*fixed_cols, *[f"{prefix}__{t}" for t in tc.expected_tables]],
+            )
+
+        if tc.json_depth > 0:
+            with conn.cursor() as res:
+                res.execute(f"SELECT table_name FROM {catalog};")
+                assert sorted([r[0] for r in res.fetchall()]) == sorted(
+                    [f"{prefix}__{t}" for t in tc.expected_tables],
+                )
+
+        for table, (cols, values) in tc.expected_values.items():
+            with conn.cursor() as res:
+                res.execute(f"SELECT {','.join(cols)} FROM {prefix}__{table};")
+                for v in values:
+                    assert res.fetchone() == v
+
+                assert res.fetchone() is None
